backend/requirements.txt
Flask==2.2.3
Flask_Cors==4.0.0
Flask_PyMongo==2.3.0
Flask_RESTful==0.3.10
flask_session==0.5.0
pymongo==3.12.3
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
